Added Some Styling

+++ server/schemas/income.js
@@ -1,34 +1,52 @@
-const { GraphQLObjectType, GraphQLString, GraphQLID, GraphQLFloat, GraphQLList, GraphQLNonNull } = require('graphql');
+const { GraphQLObjectType, GraphQLInputObjectType, GraphQLString, GraphQLID, GraphQLFloat, GraphQLList, GraphQLNonNull } = require('graphql');
+const { GraphQLDate } = require('graphql-iso-date');
 const Income = require('../models/Income');
-const User = require('../models/User');
+const UserType = require('./user').UserType;
 
 const IncomeType = new GraphQLObjectType({
   name: 'Income',
   fields: () => ({
     id: { type: GraphQLID },
     user: {
-      type: require('./user').UserType,
+      type: UserType,
       resolve(parent, args) {
-        return User.findById(parent.user);
+        return require('../models/User').findById(parent.user);
       }
     },
     amount: { type: GraphQLFloat },
     category: { type: GraphQLString },
-    date: { type: GraphQLString }
+    date: { type: GraphQLDate },
+    description: { type: GraphQLString }
   })
+});
+
+const IncomeInputType = new GraphQLInputObjectType({
+  name: 'IncomeInput',
+  fields: {
+    amount: { type: new GraphQLNonNull(GraphQLFloat) },
+    category: { type: new GraphQLNonNull(GraphQLString) },
+    date: { type: GraphQLDate },
+    description: { type: new GraphQLNonNull(GraphQLString) }
+  }
 });
 
 const incomeQueries = {
   incomes: {
     type: new GraphQLList(IncomeType),
-    resolve(parent, args) {
-      return Income.find({});
+    resolve(parent, args, context) {
+      if (!context.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
+      return Income.find({ user: context.userId });
     }
   },
   income: {
     type: IncomeType,
     args: { id: { type: GraphQLID } },
-    resolve(parent, args) {
+    resolve(parent, args, context) {
+      if (!context.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       return Income.findById(args.id);
     }
   }
@@ -38,15 +56,15 @@ const incomeMutations = {
   addIncome: {
     type: IncomeType,
     args: {
-      user: { type: new GraphQLNonNull(GraphQLID) },
-      amount: { type: new GraphQLNonNull(GraphQLFloat) },
-      category: { type: new GraphQLNonNull(GraphQLString) }
+      incomeInput: { type: new GraphQLNonNull(IncomeInputType) }
     },
-    resolve(parent, args) {
-      let income = new Income({
-        user: args.user,
-        amount: args.amount,
-        category: args.category
+    resolve(parent, { incomeInput }, context) {
+      if (!context.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
+      const income = new Income({
+        user: context.userId,
+        ...incomeInput
       });
       return income.save();
     }
@@ -56,15 +74,20 @@ const incomeMutations = {
     args: {
       id: { type: new GraphQLNonNull(GraphQLID) },
       amount: { type: GraphQLFloat },
-      category: { type: GraphQLString }
+      category: { type: GraphQLString },
+      description: { type: GraphQLString }
     },
-    resolve(parent, args) {
+    resolve(parent, args, context) {
+      if (!context.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       return Income.findByIdAndUpdate(
         args.id,
         {
           $set: {
             amount: args.amount,
-            category: args.category
+            category: args.category,
+            description: args.description
           }
         },
         { new: true }
@@ -74,7 +97,10 @@ const incomeMutations = {
   deleteIncome: {
     type: IncomeType,
     args: { id: { type: new GraphQLNonNull(GraphQLID) } },
-    resolve(parent, args) {
+    resolve(parent, args, context) {
+      if (!context.isAuth) {
+        throw new Error('Unauthenticated!');
+      }
       return Income.findByIdAndRemove(args.id);
     }
   }
@@ -82,6 +108,7 @@ const incomeMutations = {
 
 module.exports = {
   IncomeType,
+  IncomeInputType,
   incomeQueries,
   incomeMutations
 };
